#1496 moved smviomodules to sub-package smv.io
