+++ src/main/python/smv/io/__init__.py
@@ -1,0 +1,2 @@
+from smv.io.inputs import *
+from smv.io.outputs import *
