+++ src/main/python/smv/io/base.py
@@ -149,87 +149,3 @@ class SmvOutput(SmvIoModule):
             raise SmvRuntimeError("SmvOutput modules depend on a single input, more are given: {}"\
                 .format(", ".join([m.fqn() for m in self.requiresDS()]))
             )
-
-
-class SmvJdbcInputTable(SmvInput):
-    """
-        User need to implement 
-
-            - connectionName
-            - tableName
-    """
-
-    def doRun(self, known):
-        conn = self.get_connection()
-        builder = self.smvApp.sqlContext.read\
-            .format('jdbc')\
-            .option('url', conn.url)
-
-        if (conn.driver is not None):
-            builder = builder.option('driver', conn.driver)
-        if (conn.user is not None):
-            builder = builder.option('user', conn.user)
-        if (conn.password is not None):
-            builder = builder.option('password', conn.password)
-
-        return builder\
-            .option('dbtable', self.tableName())\
-            .load()
-
-
-class SmvJdbcOutputTable(SmvOutput):
-    """
-        User need to implement 
-
-            - requiresDS
-            - connectionName
-            - tableName
-            - writeMode: optional, default "errorifexists"
-    """
-
-    def writeMode(self):
-        """
-            Write mode for Spark DataFrameWriter.
-            Valid values:
-
-                - "append"
-                - "overwrite"
-                - "ignore"
-                - "error" or "errorifexists" (default)
-        """
-        return "errorifexists"
-
-
-    def doRun(self, known):
-        self.assert_single_input()
-        i = self.RunParams(known)
-
-        conn = self.get_connection()
-        data = i[self.requiresDS()[0]]
-
-        builder = data.write\
-            .format("jdbc") \
-            .mode(self.writeMode()) \
-            .option('url', conn.url)
-
-        if (conn.driver is not None):
-            builder = builder.option('driver', conn.driver)
-        if (conn.user is not None):
-            builder = builder.option('user', conn.user)
-        if (conn.password is not None):
-            builder = builder.option('password', conn.password)
-
-        builder \
-            .option("dbtable", self.tableName()) \
-            .save()
-
-        # return data back for meta calculation
-        # TODO: need to review whether we should even calculate meta for output
-        return data
-
-
-__all__ = [
-    'SmvJdbcInputTable',
-    'SmvJdbcOutputTable',
-    'SmvOutput',
-]
