+++ src/test/python/testJdbc/stage/modules.py
@@ -12,6 +12,7 @@
 # limitations under the License.
 
 from smv import *
+from smv.io import SmvJdbcInputTable, SmvJdbcOutputTable
 
 class MyJdbcTable(SmvJdbcTable):
     def tableName(self):
